move player attributes table to components folder; update usages
