+++ dashboard-ui/src/components/PlayerAttributesTable.js
