+++ dashboard-ui/src/stories/AttributeTable.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AttributeItem from '../components/AttributeItem';
-import PlayerAttributesTable from '../widgets/PlayerAttributesTable';
+import PlayerAttributesTable from '../components/PlayerAttributesTable';
 
 import { getAttributeItemData, getAttributeComparisonTableData } from './utils/storyDataGenerators';
